% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{compareEnrichBin}
\alias{compareEnrichBin}
\title{compareEnrichBin}
\usage{
compareEnrichBin(
  gene.group,
  pattern,
  bin.width = 0.2,
  stride = 0.1,
  species = "human",
  ont = "BP",
  universe = FALSE,
  ...
)
}
\arguments{
\item{gene.group}{A data frame indicating genes in each pattern}

\item{pattern}{The expression pattern for enrichment analysis}

\item{bin.width}{The width of each bin}

\item{stride}{Stride of each step}

\item{species}{Select from human or mouse}

\item{ont}{One of "BP", "MF", and "CC" subontologies, or "ALL" for all three}

\item{universe}{Pass to the universe paramenter of enrichGO}

\item{...}{Pass to the function enrichGO}
}
\value{
A compareClusterResult object summarizing enrichment results across bins
}
\description{
Bin-based enrichment for ordered genes in a specific pattern
}
\details{
This function partitions genes within a temporal expression pattern into bins
         along pseudotime and performs GO enrichment on each bin. The output contains
         enrichment results for all bins
}
\author{
Zhicheng Ji, Changxin Wan, Beijie Ji
}
