% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrichBin}
\alias{enrichBin}
\title{enrichBin}
\usage{
enrichBin(
  gene.group,
  pattern,
  bin.width = 0.2,
  stride = 0.1,
  species = "mouse",
  ont = "BP",
  ...
)
}
\arguments{
\item{gene.group}{a data frame indicate genes in each pattern}

\item{pattern}{the pattern for enrichment analysis}

\item{bin.width}{the width of each bin}

\item{stride}{stride of each step}

\item{species}{select from human or mouse}

\item{ont}{One of "BP", "MF", and "CC" subontologies, or "ALL" for all three}

\item{...}{pass to the function enrichGO}
}
\value{
A list with genes and enrichResult instance
}
\description{
Enrichment for ordered genes in specific pattern
}
\details{
This function generates a list with genes and enrichResult instance
}
\author{
Zhicheng Ji, Changxin Wan
}
