% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotEnrichGroup}
\alias{plotEnrichGroup}
\title{plotEnrichGroup}
\usage{
plotEnrichGroup(enrich_obj, terms = NULL, n = 10)
}
\arguments{
\item{enrich_obj}{An \code{enrichResult} object (from \code{enrichPattern()})
or one element of a group enrichment list}

\item{terms}{Character vector of GO IDs to display. If NULL, top \code{n} terms are shown.}

\item{n}{Number of top terms to show if \code{terms = NULL} (default: 10)}
}
\value{
A ggplot2 object of the bubble plot
}
\description{
Visualize GO enrichment results for a specific expression pattern
}
\details{
This function takes an enrichment result from \code{enrichPattern()}
         or group-enrichment list, selects user-defined GO terms (or top-ranked ones),
         and generates a bubble plot summarizing the enrichment. Gene ratio is shown on the x-axis, enriched terms on the y-axis,
         point size represents gene count, and color represents adjusted q-value.
}
\author{
Zhicheng Ji, Changxin Wan, Beijie Ji
}
