% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STIP.R
\name{PreprocessSTIP}
\alias{PreprocessSTIP}
\title{PreprocessSTIP}
\usage{
PreprocessSTIP(data, gl)
}
\arguments{
\item{data}{a single cell expression matrix or data.frame with rows as genes and columns as cells. Cells should follow the order of pseudo time}

\item{gl}{marked gene list}
}
\value{
A Heatmap-class object
}
\description{
State Transition Inference Prediction Preprocess
}
\details{
This function generates a table that performs (STIP) State Transition Inference Prediction
}
