% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{ScatterPlotSTIP}
\alias{ScatterPlotSTIP}
\title{ScatterPlotSTIP}
\usage{
ScatterPlotSTIP(data, gene, count = FALSE)
}
\arguments{
\item{data}{A fitted gene expression matrix}

\item{gene}{Plotted gene in the matrix}

\item{count}{set true if data is count matrix}
}
\value{
A ggplot object
}
\description{
Generate scatter plot for gene in fitted matrix
}
\details{
Input fitted expression, output a scatter plot
}
\author{
Zhicheng Ji, Changxin Wan
}
