% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisample_process.R
\name{SelectTopGenes}
\alias{SelectTopGenes}
\title{SelectTopGenes}
\usage{
SelectTopGenes(gene_mean_list, top = 1000)
}
\arguments{
\item{gene_mean_list}{A multi-sample result list returned by
\code{PseudotimeMSProcess()}.}

\item{top}{Number of top-ranked genes to select (default = 1000).}
}
\value{
A list with the same structure as \code{gene_mean_list},
but restricted to the selected top genes.
}
\description{
Select top-ranked genes from multi-sample Pseudotimecascade results
}
\details{
This function ranks genes based on the average significance across samples.
Specifically, q-values are transformed using \eqn{-\log_{10}(q)}, and the mean
transformed value across samples is used for ranking. The top \code{top}
genes are selected and all corresponding components (mean expression,
pattern assignment, q-values, and switch points) are subset accordingly.
}
\author{
Zhicheng Ji, Changxin Wan, Beijie Ji
}
