% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcell}
\alias{tcell}
\title{Subset of single cell T cell expression.}
\format{
A data frame with 1000 rows and 1001 columns:
}
\usage{
tcell
}
\description{
A dataset containing the expression profile of T cells
}
\keyword{datasets}
