% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrichPattern}
\alias{enrichPattern}
\title{enrichPattern}
\usage{
enrichPattern(gene.group, pattern, species = "mouse", ont = "BP", ...)
}
\arguments{
\item{gene.group}{a data frame indicate genes in each pattern}

\item{pattern}{the pattern for enrichment analysis}

\item{species}{select from human or mouse}

\item{ont}{One of "BP", "MF", and "CC" subontologies, or "ALL" for all three}

\item{...}{pass to the function enrichGO}
}
\value{
An enrichResult instance
}
\description{
Enrichment for genes in specific pattern
}
\details{
This function generates an enrichResult instance
}
\author{
Zhicheng Ji, Changxin Wan, Beijie Ji
}
