% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitData}
\alias{fitData}
\title{fitData}
\usage{
fitData(
  data,
  expr.cut = 0.1,
  expr.cut.rate = 0.05,
  pseudo.time = colnames(data),
  p.adjust.method = "BH",
  new_data = data.frame(pt = seq(1, ncol(data))),
  verbose = TRUE,
  mc.cores = 1
)
}
\arguments{
\item{data}{a single cell expression matrix with rows as genes and columns as cells.}

\item{expr.cut}{cutoff of lowerest expression}

\item{expr.cut.rate}{cutoff of cells lower than lowest expression}

\item{pseudo.time}{cells ranked according to pseudo time}

\item{p.adjust.method}{method for multiple hypothesis test}

\item{new_data}{input matrix for model prediction}

\item{verbose}{show message of running process}

\item{mc.cores}{number of cores for parallel computing}
}
\value{
A list contains scaled fitted gene expression matrix
}
\description{
Fit single cell gene expression data according to cell pseudotime
}
\details{
This function generates a fitted expression matrix of single cell RNA-seq
}
\author{
Zhicheng Ji, Changxin Wan
}
