% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STIP.R
\name{genePattern}
\alias{genePattern}
\title{genePattern}
\usage{
genePattern(x)
}
\arguments{
\item{x}{a single cell expression matrix or data.frame with rows as genes and columns as cells. Cells should follow the order of pseudo time}
}
\value{
A dataframe of state transtion pattern
}
\description{
State transition pattern of each gene
}
\details{
This function generates the state transition pattern of input gene
}
\author{
Zhicheng Ji, Changxin Wan
}
