% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotimecascade.R
\name{genePattern}
\alias{genePattern}
\title{genePattern}
\usage{
genePattern(data)
}
\arguments{
\item{data}{a single cell expression matrix or data.frame with rows as genes and columns as cells. Cells should follow the order of pseudo time}
}
\value{
A dataframe of state transtion pattern
}
\description{
State transition pattern of each gene
}
\details{
#' @title PreprocessPseudotime
#' @description Pseudotimecascade Preprocess
#' @details This function generates a table that performs Pseudotimecascade
#' @param data a single cell expression matrix or data.frame with rows as genes and columns as cells. Cells should follow the order of pseudo time
#' @param gl marked gene list
#' @return A Heatmap-class object
#' @export PreprocessPseudotime
PreprocessPseudotime <- function(data, gl){
  gene_group <- genePattern(data)
  plotdata <- data[rownames(gene_group), ]
  p <- PseudotimeHeatmap(plotdata, gl, as.matrix(gene_group)[, "pattern"])
  return(p)
}

This function generates the state transition pattern of input gene
}
\author{
Zhicheng Ji, Changxin Wan, Beijie Ji
}
