% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{PseudotimeHeatmap}
\alias{PseudotimeHeatmap}
\title{PseudotimeHeatmap}
\usage{
PseudotimeHeatmap(x, gl, annotation, ...)
}
\arguments{
\item{x}{A gene expression matrix}

\item{gl}{Marked gene list}

\item{annotation}{Annotation matrix for genes in expression matrix}

\item{...}{parameters passed to Heatmap}
}
\value{
A ComplexHeatmap object
}
\description{
Generate heatmap for Pseudotimecascade result
}
\details{
Input a gene expression matrix and annotation matrixes, output a heatmap
}
\author{
Zhicheng Ji, Changxin Wan, Beijie Ji
}
