% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{PseudotimeHeatmapMS}
\alias{PseudotimeHeatmapMS}
\title{PseudotimeHeatmapMS}
\usage{
PseudotimeHeatmapMS(x, gl, annotation, interval, ...)
}
\arguments{
\item{x}{A gene expression matrix}

\item{gl}{Marked gene list}

\item{annotation}{Annotation matrix for genes in expression matrix}

\item{interval}{A list contains zero points for each gene in each sample}

\item{...}{parameters passed to Heatmap}
}
\value{
A ComplexHeatmap object
}
\description{
Generate heatmap for multi-sample Pseudotimecascade result
}
\details{
Input a gene expression matrix and annotation matrixes, output a heatmap
}
\author{
Zhicheng Ji, Changxin Wan, Beijie Ji
}
